+++ src/application/FreischaltCode/FreischaltCodeService.py
@@ -6,10 +6,10 @@ from uuid import UUID, uuid4
 from opyoid import Injector, Module
 from rq import Retry
 
-from src.application.FreischaltCode.freischalt_code import FreischaltCodeCreateDto, FreischaltCodeDto
+from src.application.FreischaltCode.FreischaltCode import FreischaltCodeCreateDto, FreischaltCodeDto
 from src.application.FreischaltCode.Jobs.jobs import send_freischalt_code
 from src.domain.BackgroundJobs.BackgroundJobInterface import BackgroundJobInterface
-from src.domain.FreischaltCode.freischalt_code import FreischaltCode
+from src.domain.FreischaltCode.FreischaltCode import FreischaltCode
 from src.infrastructure.rq.RqModule import RqModule
 from src.infrastructure.sqlalchemy.repositories.RepositoriesModule import RepositoriesModule
 from src.infrastructure.sqlalchemy.repositories.FreischaltCodeRepository import FreischaltCodeRepository
