+++ src/domain/Shared/BaseDomainModel.py
