+++ src/domain/TaxDeclaration/TaxDeclaration.py
@@ -4,7 +4,7 @@ from uuid import UUID
 from pydantic import BaseModel
 
 from src.domain.Repositories.BaseRepositoryInterface import BaseRepositoryInterface
-from src.domain.Shared.base_domain_model import BaseDomainModel
+from src.domain.Shared.BaseDomainModel import BaseDomainModel
 from src.domain.Shared.status import Status
