+++ src/infrastructure/sqlalchemy/TaxDeclaration.py
@@ -3,8 +3,8 @@ from sqlalchemy import MetaData, Column, String
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.ext.declarative import declarative_base
 
-from src.domain.TaxDeclaration.tax_declaration import Status
-from src.infrastructure.sqlalchemy.base_entity import AuditedSchemaMixin
+from src.domain.TaxDeclaration.TaxDeclaration import Status
+from src.infrastructure.sqlalchemy.BaseSchema import AuditedSchemaMixin
 
 metadata = MetaData()
 BaseDbSchema = declarative_base()
