+++ src/infrastructure/sqlalchemy/repositories/base_repository.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel
 from sqlalchemy.orm import Session
 
 from src.domain.Repositories.BaseRepositoryInterface import BaseRepositoryInterface
-from src.domain.Shared.base_domain_model import BaseDomainModel
+from src.domain.Shared.BaseDomainModel import BaseDomainModel
 from src.infrastructure.sqlalchemy.FreischaltCodeSchema import BaseDbSchema
 
 T = TypeVar('T', bound=BaseDomainModel)
